soc: seperate sustainer components into files
